Fix problem where -1 gets converted to 65536

+++ src/main.cpp
@@ -78,7 +78,7 @@ int main ( int argc, char** argv )
     bool         bCustomPortNumberGiven      = false;
     int          iNumServerChannels          = DEFAULT_USED_NUM_CHANNELS;
     quint16      iPortNumber                 = DEFAULT_PORT_NUMBER;
-    quint16      iJsonRpcPortNumber          = -1;
+    int          iJsonRpcPortNumber          = -1;
     quint16      iQosNumber                  = DEFAULT_QOS_NUMBER;
     ELicenceType eLicenceType                = LT_NO_LICENCE;
     QString      strMIDISetup                = "";
